introduce connection options builder and make connection options immutable.
